snt_worldpop_extract/requirements.txt
openhexa.toolbox @ git+https://github.com/BLSQ/openhexa-toolbox@main
snt_lib @ git+https://git@github.com/BLSQ/snt_utils.git
rasterstats